--- apps/meteor/client/views/room/contextualBar/VideoConference/VideoConfPopups/VideoConfPopup/StartCallPopup/index.ts
@@ -1,1 +1,0 @@
-export { default } from './StartCallPopup';
