+++ packages/ui-video-conf/src/VideoConfPopup/VideoConfPopupInfo.tsx
@@ -1,0 +1,24 @@
+import { Box } from '@rocket.chat/fuselage';
+import type { ReactElement, ReactNode } from 'react';
+
+type VideoConfPopupInfoProps = {
+	avatar?: ReactElement;
+	icon?: ReactNode;
+	children: ReactNode;
+};
+
+const VideoConfPopupInfo = ({ avatar, icon, children }: VideoConfPopupInfoProps): ReactElement => (
+	<Box display='flex' alignItems='center'>
+		{avatar}
+		{(icon || children) && (
+			<Box display='flex' alignItems='center' mis='x8' withTruncatedText>
+				{icon}
+				<Box mis='x8' fontScale='h4' withTruncatedText>
+					{children}
+				</Box>
+			</Box>
+		)}
+	</Box>
+);
+
+export default VideoConfPopupInfo;
